Add send max balance button (#289)

+++ BankWallet/BankWallet/Models/CurrencyValue.swift
@@ -1,4 +1,6 @@
+import Foundation
+
 struct CurrencyValue {
     let currency: Currency
-    let value: Double
+    let value: Decimal
 }
